Particles added on home page
